Update README with exercise details
The README has been updated to include a more detailed description of the PHP exercises. It now specifies not only that the exercises are part of the SPAI Locarno series, but it also includes a list of the completed exercises and their versions. This addition will make it easier for others to understand the contents of this repository.